+++ probables/blooms/basebloom.py
@@ -7,9 +7,11 @@ from __future__ import (unicode_literals, absolute_import, print_function,
 import sys
 import math
 import os
+from numbers import Number
 from abc import (abstractmethod)
 from struct import (pack, unpack, calcsize, Struct)
 from binascii import (hexlify, unhexlify)
+
 from .. exceptions import (InitializationError)
 from .. hashes import (default_fnv_1a)
 from .. utilities import (is_hex_string, is_valid_file)
@@ -155,8 +157,7 @@ class BaseBloom(object):
                 depth (int): Number of permutations of the hash to generate; \
                 if None, generate `number_hashes`
             Returns:
-                List(int): A list of the hashes for the key in int form
-        '''
+                List(int): A list of the hashes for the key in int form '''
         tmp = depth if depth is not None else self.number_hashes
         return self.__hash_func(key, tmp)
 
@@ -169,6 +170,17 @@ class BaseBloom(object):
         else:
             tmp_hash = hash_function
 
+        valid_prms = (isinstance(estimated_elements, Number) and
+                      estimated_elements > 0)
+        if not valid_prms:
+            msg = 'Bloom: estimated elements must be greater than 0'
+            raise InitializationError(msg)
+        valid_prms = (isinstance(false_positive_rate, Number) and
+                      0.0 <= false_positive_rate < 1.0)
+        if not valid_prms:
+            msg = 'Bloom: false positive rate must be between 0.0 and 1.0'
+            raise InitializationError(msg)
+
         fpr = pack('f', float(false_positive_rate))
         t_fpr = unpack('f', fpr)[0]  # to mimic the c version!
         # optimal caluclations
@@ -176,6 +188,10 @@ class BaseBloom(object):
         fpr = float(false_positive_rate)
         m_bt = math.ceil((-n_els * math.log(fpr)) / 0.4804530139182)  # ln(2)^2
         number_hashes = int(round(math.log(2.0) * m_bt / n_els))
+
+        if number_hashes <= 0:  # this should never happen...
+            msg = 'Bloom: Number hashes is zero; unusable parameters provided'
+            raise InitializationError(msg)
 
         return tmp_hash, t_fpr, number_hashes, int(m_bt)
 
@@ -227,8 +243,7 @@ class BaseBloom(object):
         ''' Export the Bloom Filter as a hex string
 
             Return:
-                str: Hex representation of the Bloom Filter
-        '''
+                str: Hex representation of the Bloom Filter '''
         mybytes = pack('>QQf', self.estimated_elements,
                        self.elements_added, self.false_positive_rate)
         if self.__blm_type in ['regular', 'reg-ondisk']:
@@ -247,8 +262,7 @@ class BaseBloom(object):
 
             Args:
                 filename (str): The filename to which the Bloom Filter will \
-                be written.
-        '''
+                be written. '''
         with open(filename, 'wb') as filepointer:
             rep = self.__impt_type * self.bloom_length
             filepointer.write(pack(rep, *self.bloom))
@@ -260,8 +274,7 @@ class BaseBloom(object):
         ''' Calculate the size of the bloom on disk
 
             Returns:
-                int: Size of the Bloom Filter when exported to disk
-        '''
+                int: Size of the Bloom Filter when exported to disk '''
         tmp_b = calcsize(self.__impt_type)
         return (self.bloom_length * tmp_b) + calcsize('QQf')
 
@@ -269,8 +282,7 @@ class BaseBloom(object):
         ''' Calculate the current false positive rate based on elements added
 
             Return:
-                float: The current false positive rate
-        '''
+                float: The current false positive rate '''
         num = self.number_hashes * -1 * self.elements_added
         dbl = num / float(self.number_bits)
         exp = math.exp(dbl)
@@ -280,8 +292,7 @@ class BaseBloom(object):
         ''' Estimate the number of unique elements added
 
             Returns:
-                int: Number of elements estimated to be inserted
-        '''
+                int: Number of elements estimated to be inserted '''
         setbits = self._cnt_number_bits_set()
         log_n = math.log(1 - (float(setbits) / float(self.number_bits)))
         tmp = float(self.number_bits) / float(self.number_hashes)
@@ -312,8 +323,7 @@ class BaseBloom(object):
         ''' Add the key to the Bloom Filter
 
             Args:
-                key (str): The element to be inserted
-        '''
+                key (str): The element to be inserted '''
         hashes = self.hashes(key)
         self.add_alt(hashes)
 
@@ -322,8 +332,7 @@ class BaseBloom(object):
 
             Args:
                 hashes (list): A list of integers representing the key to \
-                insert
-        '''
+                insert '''
         for i in list(range(0, self.number_hashes)):
             k = int(hashes[i]) % self.number_bits
             idx = k // 8
@@ -338,8 +347,7 @@ class BaseBloom(object):
             Args:
                 key (str): The element to be checked
             Returns:
-                bool: True if likely encountered, False if definately not
-        '''
+                bool: True if likely encountered, False if definately not '''
         hashes = self.hashes(key)
         return self.check_alt(hashes)
 
@@ -350,8 +358,7 @@ class BaseBloom(object):
                 hashes (list): A list of integers representing the key to \
                 check
             Returns:
-                bool: True if likely encountered, False if definately not
-        '''
+                bool: True if likely encountered, False if definately not '''
         for i in list(range(0, self.number_hashes)):
             k = int(hashes[i]) % self.number_bits
             if (int(self._get_element(k // 8)) & int((1 << (k % 8)))) == 0: