+++ probables/blooms/bloom.py
@@ -10,6 +10,7 @@ import os
 import mmap
 from struct import (pack, unpack, calcsize)
 from shutil import (copyfile)
+
 from . basebloom import (BaseBloom)
 from .. exceptions import (InitializationError, NotSupportedError)
 from .. utilities import (is_hex_string, is_valid_file)
@@ -133,8 +134,7 @@ class BloomFilter(BaseBloom):
                 TypeError: When second is not either a :class:`BloomFilter` \
                 or :class:`BloomFilterOnDisk`
             Note:
-                `second` may be a BloomFilterOnDisk object
-        '''
+                `second` may be a BloomFilterOnDisk object '''
         if not _verify_not_type_mismatch(second):
             raise TypeError(MISMATCH_MSG)
 
@@ -155,8 +155,7 @@ class BloomFilter(BaseBloom):
                 TypeError: When second is not either a :class:`BloomFilter` \
                 or :class:`BloomFilterOnDisk`
             Note:
-                `second` may be a BloomFilterOnDisk object
-        '''
+                `second` may be a BloomFilterOnDisk object '''
         if not _verify_not_type_mismatch(second):
             raise TypeError(MISMATCH_MSG)
 
@@ -177,8 +176,7 @@ class BloomFilter(BaseBloom):
                 TypeError: When second is not either a :class:`BloomFilter` \
                 or :class:`BloomFilterOnDisk`
             Note:
-                `second` may be a BloomFilterOnDisk object
-        '''
+                `second` may be a BloomFilterOnDisk object '''
         if not _verify_not_type_mismatch(second):
             raise TypeError(MISMATCH_MSG)
 
@@ -298,8 +296,7 @@ class BloomFilterOnDisk(BaseBloom):
                 filename (str): The filename to which the Bloom Filter will \
                 be exported
             Note:
-                Only exported if the filename is not the original filename
-        '''
+                Only exported if the filename is not the original filename '''
         self.__update()
         if filename != self.__filename:
             # setup the new bloom filter
@@ -322,8 +319,7 @@ class BloomFilterOnDisk(BaseBloom):
                 TypeError: When second is not either a :class:`BloomFilter` \
                 or :class:`BloomFilterOnDisk`
             Note:
-                `second` may be a BloomFilter object
-        '''
+                `second` may be a BloomFilter object '''
         if not _verify_not_type_mismatch(second):
             raise TypeError(MISMATCH_MSG)
 
@@ -346,8 +342,7 @@ class BloomFilterOnDisk(BaseBloom):
                 TypeError: When second is not either a :class:`BloomFilter` \
                 or :class:`BloomFilterOnDisk`
             Note:
-                `second` may be a BloomFilter object
-        '''
+                `second` may be a BloomFilter object '''
         if not _verify_not_type_mismatch(second):
             raise TypeError(MISMATCH_MSG)
 
@@ -369,8 +364,7 @@ class BloomFilterOnDisk(BaseBloom):
                 TypeError: When second is not either a :class:`BloomFilter` \
                 or :class:`BloomFilterOnDisk`
             Note:
-                `second` may be a BloomFilter object
-        '''
+                `second` may be a BloomFilter object '''
         if not _verify_not_type_mismatch(second):
             raise TypeError(MISMATCH_MSG)
 
@@ -384,8 +378,7 @@ class BloomFilterOnDisk(BaseBloom):
 
             Raises:
                 NotSupportedError: This functionality is currently not \
-                supported
-        '''
+                supported '''
         msg = ('`export_hex` is currently not supported by the on disk '
                'Bloom Filter')
         raise NotSupportedError(msg)