+++ probables/countminsketch/countminsketch.py
@@ -7,7 +7,9 @@ from __future__ import (unicode_literals, absolute_import, print_function,
                         division)
 import os
 import math
+from numbers import Number
 from struct import (pack, unpack, calcsize)
+
 from .. exceptions import (InitializationError, NotSupportedError)
 from .. hashes import (default_fnv_1a)
 from .. utilities import (is_valid_file)
@@ -55,18 +57,28 @@ class CountMinSketch(object):
         if is_valid_file(filepath):
             self.__load(filepath)
         elif width is not None and depth is not None:
-            self.__width = width
-            self.__depth = depth
-            self.__confidence = 1 - (1 / math.pow(2, depth))
-            self.__error_rate = 2 / width
-            self._bins = [0] * (self.__width * self.__depth)
+            valid_prms = (isinstance(width, Number) and width > 0 and
+                          isinstance(depth, Number) and depth > 0)
+            if not valid_prms:
+                msg = 'CountMinSketch: width and depth must be greater than 0'
+                raise InitializationError(msg)
+            self.__width = int(width)
+            self.__depth = int(depth)
+            self.__confidence = 1 - (1 / math.pow(2, self.depth))
+            self.__error_rate = 2 / self.width
+            self._bins = [0] * (self.width * self.depth)
         elif confidence is not None and error_rate is not None:
+            valid_prms = (isinstance(confidence, Number) and confidence > 0 and
+                          isinstance(error_rate, Number) and error_rate > 0)
+            if not valid_prms:
+                msg = 'CountMinSketch: width and depth must be greater than 0'
+                raise InitializationError(msg)
             self.__confidence = confidence
             self.__error_rate = error_rate
             self.__width = math.ceil(2 / error_rate)
             numerator = (-1 * math.log(1 - confidence))
             self.__depth = math.ceil(numerator / 0.6931471805599453)
-            self._bins = [0] * int(self.__width * self.__depth)
+            self._bins = [0] * int(self.width * self.depth)
         else:
             msg = ('Must provide one of the following to initialize the '
                    'Count-Min Sketch:\n'
@@ -139,8 +151,7 @@ class CountMinSketch(object):
                 Valid values:
                     * 'min' or None
                     * 'mean'
-                    * 'mean-min'
-        '''
+                    * 'mean-min' '''
         if self.__query_method == self.__min_query:
             return 'min'
         elif self.__query_method == self.__mean_query:
@@ -180,9 +191,8 @@ class CountMinSketch(object):
                 if None, generate `number_hashes`
 
             Returns:
-                List(int): A list of the hashes for the key in int form
-        '''
-        t_depth = self.__depth if depth is None else depth
+                List(int): A list of the hashes for the key in int form '''
+        t_depth = self.depth if depth is None else depth
         return self._hash_function(key, t_depth)
 
     def add(self, key, num_els=1):
@@ -193,8 +203,7 @@ class CountMinSketch(object):
                 num_els (int): The number of times to insert the element
             Returns:
                 int: The number of times the element was likely inserted \
-                after the insertion
-        '''
+                after the insertion '''
         hashes = self.hashes(key)
         return self.add_alt(hashes, num_els)
 
@@ -206,18 +215,17 @@ class CountMinSketch(object):
                 num_els (int): The number of times to insert the element
             Returns:
                 int: The number of times the element was likely inserted \
-                after the insertion
-        '''
+                after the insertion '''
         res = list()
         for i, val in enumerate(hashes):
-            t_bin = (val % self.__width) + (i * self.__width)
+            t_bin = (val % self.width) + (i * self.width)
             self._bins[t_bin] += num_els
             if self._bins[t_bin] > INT32_T_MAX:
                 self._bins[t_bin] = INT32_T_MAX
             res.append(self._bins[t_bin])
         self.__elements_added += num_els
 
-        if self.__elements_added > INT64_T_MAX:
+        if self.elements_added > INT64_T_MAX:
             self.__elements_added = INT64_T_MAX
         return self.__query_method(sorted(res))
 
@@ -229,8 +237,7 @@ class CountMinSketch(object):
                 num_els (int): The number of times to remove the element
             Returns:
                 int: The number of times the element was likely inserted \
-                after the removal
-        '''
+                after the removal '''
         hashes = self.hashes(key)
         return self.remove_alt(hashes, num_els)
 
@@ -242,17 +249,16 @@ class CountMinSketch(object):
                 num_els (int): The number of times to remove the element
             Returns:
                 int: The number of times the element was likely inserted \
-                after the removal
-        '''
+                after the removal '''
         res = list()
         for i, val in enumerate(hashes):
-            t_bin = (val % self.__width) + (i * self.__width)
+            t_bin = (val % self.width) + (i * self.width)
             self._bins[t_bin] -= num_els
             if self._bins[t_bin] < INT32_T_MIN:
                 self._bins[t_bin] = INT32_T_MIN
             res.append(self._bins[t_bin])
         self.__elements_added -= num_els
-        if self.__elements_added < INT64_T_MIN:
+        if self.elements_added < INT64_T_MIN:
             self.__elements_added = INT64_T_MIN
 
         return self.__query_method(sorted(res))
@@ -263,8 +269,7 @@ class CountMinSketch(object):
             Args:
                 key (str): The key to check the number of times inserted
             Returns:
-                int: The number of times the element was likely inserted
-        '''
+                int: The number of times the element was likely inserted '''
         hashes = self.hashes(key)
         return self.check_alt(hashes)
 
@@ -275,8 +280,7 @@ class CountMinSketch(object):
             Args:
                 hashes (list): The hashes representing the element to check
             Returns:
-                int: The number of times the element was likely inserted
-        '''
+                int: The number of times the element was likely inserted '''
         bins = self.__get_values_sorted(hashes)
         return self.__query_method(bins)
 
@@ -290,8 +294,8 @@ class CountMinSketch(object):
             # write out the bins
             rep = 'i' * len(self._bins)
             filepointer.write(pack(rep, *self._bins))
-            filepointer.write(pack('IIq', self.__width, self.__depth,
-                                   self.__elements_added))
+            filepointer.write(pack('IIq', self.width, self.depth,
+                                   self.elements_added))
 
     def __load(self, filepath):
         ''' load the count-min sketch from file '''
@@ -302,11 +306,11 @@ class CountMinSketch(object):
             self.__width = mybytes[0]
             self.__depth = mybytes[1]
             self.__elements_added = mybytes[2]
-            self.__confidence = 1 - (1 / math.pow(2, self.__depth))
-            self.__error_rate = 2 / self.__width
+            self.__confidence = 1 - (1 / math.pow(2, self.depth))
+            self.__error_rate = 2 / self.width
 
             filepointer.seek(0, os.SEEK_SET)
-            length = self.__width * self.__depth
+            length = self.width * self.depth
             rep = 'i' * length
             offset = calcsize(rep)
             self._bins = list(unpack(rep, filepointer.read(offset)))
@@ -315,7 +319,7 @@ class CountMinSketch(object):
         ''' get the values sorted '''
         bins = list()
         for i, val in enumerate(hashes):
-            t_bin = (val % self.__width) + (i * self.__width)
+            t_bin = (val % self.width) + (i * self.width)
             bins.append(self._bins[t_bin])
         bins.sort()
         return bins
@@ -327,21 +331,21 @@ class CountMinSketch(object):
 
     def __mean_query(self, results):
         ''' generate the mean query; assumes sorted list '''
-        return sum(results) // self.__depth
+        return sum(results) // self.depth
 
     def __mean_min_query(self, results):
         ''' generate the mean-min query; assumes sorted list '''
         meanmin = list()
         for t_bin in results:
-            diff = self.__elements_added - t_bin
-            calc = t_bin - diff // (self.__width - 1)
+            diff = self.elements_added - t_bin
+            calc = t_bin - diff // (self.width - 1)
             meanmin.append(calc)
         meanmin.sort()
-        if self.__depth % 2 == 0:
-            calc = meanmin[self.__depth//2] + meanmin[self.__depth//2 - 1]
+        if self.depth % 2 == 0:
+            calc = meanmin[self.depth//2] + meanmin[self.depth//2 - 1]
             res = calc // 2
         else:
-            res = meanmin[self.__depth//2]
+            res = meanmin[self.depth//2]
         return res
 
 
@@ -482,8 +486,7 @@ class HeavyHitters(CountMinSketch):
             Returns:
                 int: Number of times key has been inserted
             Note:
-                Override function
-        '''
+                Override function '''
         hashes = self.hashes(key)
         return self.add_alt(key, hashes, num_els)
 
@@ -588,8 +591,7 @@ class StreamThreshold(CountMinSketch):
             Returns:
                 int: Number of times key has been inserted
             Note:
-                Override function
-        '''
+                Override function '''
         hashes = self.hashes(key)
         return self.add_alt(key, hashes, num_els)