+++ probables/cuckoo/cuckoo.py
@@ -7,11 +7,12 @@ from __future__ import (unicode_literals, absolute_import, print_function,
                         division)
 import os
 import random
+from numbers import Number
 from struct import (pack, unpack, calcsize)
 
 from .. hashes import (fnv_1a)
-from .. utilities import (get_x_bits)
-from .. exceptions import (CuckooFilterFullError)
+from .. utilities import (is_valid_file, get_x_bits)
+from .. exceptions import (CuckooFilterFullError, InitializationError)
 
 
 class CuckooFilter(object):
@@ -23,29 +24,49 @@ class CuckooFilter(object):
             max_swaps (int): The number of cuckoo swaps before stopping
             expansion_rate (int): The rate at which to expand
             auto_expand (bool): If the filter should automatically expand
+            finger_size (int): The size of the fingerprint to use in bytes \
+            (between 1 and 4); exported as 4 bytes; up to the user to reset \
+            the size correctly on import
             filename (str): The path to the file to load or None if no file
+            hash_function (function): Hashing strategy function to use \
+            `hf(key)`
         Returns:
-            CuckooFilter: A Cuckoo Filter object
-    '''
+            CuckooFilter: A Cuckoo Filter object '''
     def __init__(self, capacity=10000, bucket_size=4, max_swaps=500,
-                 expansion_rate=2, auto_expand=True, filepath=None):
+                 expansion_rate=2, auto_expand=True, finger_size=4,
+                 filepath=None, hash_function=None):
         ''' setup the data structure '''
-        self._bucket_size = bucket_size
-        self._cuckoo_capacity = capacity
-        self.__max_cuckoo_swaps = max_swaps
+        valid_prms = (isinstance(capacity, Number) and capacity >= 1 and
+                      isinstance(bucket_size, Number) and bucket_size >= 1 and
+                      isinstance(max_swaps, Number) and max_swaps >= 1)
+        if not valid_prms:
+            msg = ('CuckooFilter: capacity, bucket_size, and max_swaps '
+                   'must be an integer greater than 0')
+            raise InitializationError(msg)
+        self._bucket_size = int(bucket_size)
+        self._cuckoo_capacity = int(capacity)
+        self.__max_cuckoo_swaps = int(max_swaps)
         self.__expansion_rate = None
         self.expansion_rate = expansion_rate
         self.__auto_expand = None
         self.auto_expand = auto_expand
+        self.__fingerprint_size = None
+        self.fingerprint_size = finger_size
 
-        self.__hash_func = fnv_1a
+        if hash_function is None:
+            self.__hash_func = fnv_1a
+        else:
+            self.__hash_func = hash_function
         self._inserted_elements = 0
         if filepath is None:
             self._buckets = list()
             for _ in range(self.capacity):
-                self._buckets.append(list())
-        else:
+                self.buckets.append(list())
+        elif is_valid_file(filepath):
             self._load(filepath)
+        else:
+            msg = 'CuckooFilter: failed to load provided file'
+            raise InitializationError(msg)
 
     def __contains__(self, key):
         ''' setup the `in` keyword '''
@@ -109,7 +130,7 @@ class CuckooFilter(object):
 
     @property
     def expansion_rate(self):
-        ''' int: The rate at expansion when the filter grows'''
+        ''' int: The rate at expansion when the filter grows '''
         return self.__expansion_rate
 
     @expansion_rate.setter
@@ -126,6 +147,26 @@ class CuckooFilter(object):
     def auto_expand(self, val):
         ''' set the self expand value '''
         self.__auto_expand = bool(val)
+
+    @property
+    def fingerprint_size(self):
+        ''' int: The size in bytes of the fingerprint
+
+            Raises:
+                ValueError: If the size is not between 1 and 4
+            Note:
+                The size of the fingerprint must be between 1 and 4 '''
+        return self.__fingerprint_size // 8
+
+    @fingerprint_size.setter
+    def fingerprint_size(self, val):
+        ''' set the fingerprint size '''
+        tmp = int(val)
+        if not 1 <= tmp <= 4:
+            msg = ('{}: fingerprint size must be between 1 '
+                   'and 4').format(self.__class__.__name__)
+            raise ValueError(msg)
+        self.__fingerprint_size = tmp * 8  # bytes to bits
 
     def load_factor(self):
         ''' float: How full the Cuckoo Filter is currently '''
@@ -145,18 +186,15 @@ class CuckooFilter(object):
         if is_present is not None:  # already there, nothing to do
             return
         finger = self._insert_fingerprint(fingerprint, idx_1, idx_2)
-        if finger is None:
-            return
-        elif self.auto_expand:
-            self.__expand_logic(finger)
-        else:
-            raise CuckooFilterFullError('The CuckooFilter is currently full')
+        self._deal_with_insertion(finger)
 
     def check(self, key):
         ''' Check if an element is in the filter
 
             Args:
-                key (str): Element to check '''
+                key (str): Element to check
+            Returns:
+                bool: True if likely present, False if definately not '''
         idx_1, idx_2, fingerprint = self._generate_fingerprint_info(key)
         is_present = self._check_if_present(idx_1, idx_2, fingerprint)
         if is_present is not None:
@@ -167,7 +205,9 @@ class CuckooFilter(object):
         ''' Remove an element from the filter
 
             Args:
-                key (str): Element to remove '''
+                key (str): Element to remove
+            Returns:
+                bool: True if removed, False if not present '''
         idx_1, idx_2, fingerprint = self._generate_fingerprint_info(key)
         idx = self._check_if_present(idx_1, idx_2, fingerprint)
         if idx is None:
@@ -180,8 +220,7 @@ class CuckooFilter(object):
         ''' Export cuckoo filter to file
 
             Args:
-                filename (str): Path to file to export
-        '''
+                filename (str): Path to file to export '''
         with open(filename, 'wb') as filepointer:
             for bucket in self.buckets:
                 # do something for each...
@@ -195,7 +234,7 @@ class CuckooFilter(object):
 
     def expand(self):
         ''' Expand the cuckoo filter '''
-        self.__expand_logic(None)
+        self._expand_logic(None)
 
     def _insert_fingerprint(self, fingerprint, idx_1, idx_2):
         ''' insert a fingerprint '''
@@ -246,11 +285,11 @@ class CuckooFilter(object):
             filepointer.seek(0, os.SEEK_SET)
             self._buckets = list()
             for i in range(self.capacity):
-                self._buckets.append(list())
+                self.buckets.append(list())
                 for _ in range(self.bucket_size):
                     fingerprint = unpack('I', filepointer.read(int_size))[0]
                     if fingerprint != 0:
-                        self._buckets[i].append(fingerprint)
+                        self.buckets[i].append(fingerprint)
                         self._inserted_elements += 1
 
     def _check_if_present(self, idx_1, idx_2, fingerprint):
@@ -268,7 +307,7 @@ class CuckooFilter(object):
             return True
         return False
 
-    def __expand_logic(self, extra_fingerprint):
+    def _expand_logic(self, extra_fingerprint):
         ''' the logic to acutally expand the cuckoo filter '''
         # get all the fingerprints
         fingerprints = self._setup_expand(extra_fingerprint)
@@ -314,7 +353,7 @@ class CuckooFilter(object):
         '''
         # generate the fingerprint along with the two possible indecies
         hash_val = self.__hash_func(key)
-        fingerprint = get_x_bits(hash_val, 64, 32, True)
+        fingerprint = get_x_bits(hash_val, 64, self.__fingerprint_size, True)
         idx_1, idx_2 = self._indicies_from_fingerprint(fingerprint)
 
         # NOTE: This should never happen...
@@ -322,3 +361,13 @@ class CuckooFilter(object):
             msg = ('Either idx_1 {0} or idx_2 {1} is greater than {2}')
             raise ValueError(msg.format(idx_1, idx_2, self.capacity))
         return idx_1, idx_2, fingerprint
+
+    def _deal_with_insertion(self, finger):
+        ''' some code to handle the insertion the same '''
+        if finger is None:
+            return
+        elif self.auto_expand:
+            self._expand_logic(finger)
+        else:
+            msg = 'The {} is currently full'.format(self.__class__.__name__)
+            raise CuckooFilterFullError(msg)