+++ probables/exceptions.py
@@ -6,8 +6,7 @@ class ProbablesBaseException(Exception):
     ''' Base ProbablesBaseException
 
         Args:
-            message (str): The error message to be reported
-    '''
+            message (str): The error message to be reported '''
     def __init__(self, message):
         self.message = message
         super(ProbablesBaseException, self).__init__(self.message)
@@ -23,8 +22,7 @@ class InitializationError(ProbablesBaseException):
     ''' Initialization Exception
 
         Args:
-            message (str): The initialization error messge
-    '''
+            message (str): The initialization error messge '''
     def __init__(self, message):
         self.message = message
         super(InitializationError, self).__init__(self.message)
@@ -34,8 +32,7 @@ class NotSupportedError(ProbablesBaseException):
     ''' Not Supported Functionality Exception
 
         Args:
-            message (str): The error message to be reported
-    '''
+            message (str): The error message to be reported '''
     def __init__(self, message):
         self.message = message
         super(NotSupportedError, self).__init__(self.message)
@@ -45,8 +42,7 @@ class CuckooFilterFullError(ProbablesBaseException):
     ''' Cuckoo Filter Full Exception
 
         Args:
-            message (str): The error message to be reported
-    '''
+            message (str): The error message to be reported '''
     def __init__(self, message):
         self.message = message
         super(CuckooFilterFullError, self).__init__(self.message)