+++ tests/test_utilities.py
@@ -1,9 +1,10 @@
+''' probables utilitites tests '''
 from __future__ import (unicode_literals, absolute_import, print_function)
 import unittest
 import os
 
-from . utilities import (different_hash)
 from probables.utilities import (is_hex_string, is_valid_file, get_x_bits)
+from . utilities import (different_hash)
 
 
 class TestProbablesUtilities(unittest.TestCase):
@@ -42,32 +43,32 @@ class TestProbablesUtilities(unittest.TestCase):
         ''' test it on much larger numbers '''
         res = different_hash('this is a test', 1)[0]
         # 1010100101011011100100010101010011110000001010011010000101001011
-        t1 = get_x_bits(res, 64, 32, True)
-        t2 = get_x_bits(res, 64, 32, False)
-        self.assertEqual(4029260107, t1)
-        self.assertEqual(2841350484, t2)
+        tmp1 = get_x_bits(res, 64, 32, True)
+        tmp2 = get_x_bits(res, 64, 32, False)
+        self.assertEqual(4029260107, tmp1)
+        self.assertEqual(2841350484, tmp2)
 
-        t1 = get_x_bits(res, 64, 16, True)
-        t2 = get_x_bits(res, 64, 16, False)
-        self.assertEqual(41291, t1)
-        self.assertEqual(43355, t2)
+        tmp1 = get_x_bits(res, 64, 16, True)
+        tmp2 = get_x_bits(res, 64, 16, False)
+        self.assertEqual(41291, tmp1)
+        self.assertEqual(43355, tmp2)
 
-        t1 = get_x_bits(res, 64, 8, True)
-        t2 = get_x_bits(res, 64, 8, False)
-        self.assertEqual(75, t1)
-        self.assertEqual(169, t2)
+        tmp1 = get_x_bits(res, 64, 8, True)
+        tmp2 = get_x_bits(res, 64, 8, False)
+        self.assertEqual(75, tmp1)
+        self.assertEqual(169, tmp2)
 
-        t1 = get_x_bits(res, 64, 4, True)
-        t2 = get_x_bits(res, 64, 4, False)
-        self.assertEqual(11, t1)
-        self.assertEqual(10, t2)
+        tmp1 = get_x_bits(res, 64, 4, True)
+        tmp2 = get_x_bits(res, 64, 4, False)
+        self.assertEqual(11, tmp1)
+        self.assertEqual(10, tmp2)
 
-        t1 = get_x_bits(res, 64, 2, True)
-        t2 = get_x_bits(res, 64, 2, False)
-        self.assertEqual(3, t1)
-        self.assertEqual(2, t2)
+        tmp1 = get_x_bits(res, 64, 2, True)
+        tmp2 = get_x_bits(res, 64, 2, False)
+        self.assertEqual(3, tmp1)
+        self.assertEqual(2, tmp2)
 
-        t1 = get_x_bits(res, 64, 1, True)
-        t2 = get_x_bits(res, 64, 1, False)
-        self.assertEqual(1, t1)
-        self.assertEqual(1, t2)
+        tmp1 = get_x_bits(res, 64, 1, True)
+        tmp2 = get_x_bits(res, 64, 1, False)
+        self.assertEqual(1, tmp1)
+        self.assertEqual(1, tmp2)